Imports results from Tau-Argus

Imports into R the results generated by Tau-Argus from the information contained in an arb file.
(Importe dans R les résultats générés par Tau-Argus à partir des informations contenues dans un fichier arb.)

Utilisation

import(arb_filename)

Arguments

arb_filename

name of the arb file (with extension) containing the information needed for the import.
(nom du fichier arb (avec extension) contenant les informations nécessaires à l'import.)

Valeur de retour

A list of one or more data.frames. Each data.frame corresponds to to the result of a tabulation. The names of the tables filled in the lines of the batch of the form // <TABLE_ID> "..." are recovered.
(Une liste d'un ou plusieurs data.frames. Chaque data.frame correspond au résultat d'une tabulation. Les noms des tableaux renseignés dans les lignes du batch de la forme // <TABLE_ID> "..." sont récupérés.)

Détails

Requires that the batch has been executed and finished without error. In order to import immediately after the batch has been executed, this function will be most often called via link{run_arb} (by setting import = TRUE).

It is only possible (for the moment) to import results of type "2" (csv for pivot-table) and "4" (sbs). If it is not possible to import for a given tabulation, an empty data.frame is returned (with a message).

(Nécessite que le batch ait été exécuté et se soit terminé sans erreur. Afin d'importer immédiatement après exécution du batch, cette fonction sera ainsi le plus souvent appelée via run_arb() (en paramétrant import = TRUE).

Il n'est possible (pour l'instant) que d'importer les résultats de type "2" (csv for pivot-table) et "4" (sbs). En cas d'impossibilité de l'import pour une tabulation donnée, un data.frame vide est retourné (avec un message d'avertissement).)

Attributes

Each data.frame is associated with a set of attributes (metadata) allowing to keep a trace of the specifications passed to Tau-Argus.

Attributes systematically present : explanatory_vars, response_var, safetyrule, suppress, linked, output_type.

Attributes present only if the corresponding option has been filled in by the user: shadow_var, cost_var, output_options.

(À chaque data.frame est associé un ensemble d'attributs (métadonnées) permettant de conserver une trace des spécifications passées à Tau-Argus.

Attributs systématiquement présents : explanatory_vars, response_var, safetyrule, suppress, linked, output_type.

Attributs présents uniquement si l'option correspondante a été renseignée par l'utilisateur : shadow_var, cost_var, output_options.)

See also

The function micro_rtauargus(), which uses this function and inherits its parameters.
(La fonction micro_rtauargus(), qui utilise cette fonction et hérite de ses paramètres.)