Creates a batch file (.arb) for tabular data in order to run Tau Argus

Function doesn't check if the asc and rda files exists
(Crée un fichier batch (.arb) pour données tabulées exécutable par Tau-Argus en ligne de commande. La fonction ne vérifie pas si les fichiers asc et rda existent.)

Usage

tab_arb(
  arb_filename = NULL,
  tab_filename,
  rda_filename = NULL,
  hst_filename = NULL,
  explanatory_vars,
  value = getOption("rtauargus.response_var"),
  safety_rules,
  suppress,
  output_names = NULL,
  output_type = getOption("rtauargus.output_type"),
  output_options = getOption("rtauargus.output_options"),
  gointeractive = FALSE
)

Arguments

arb_filename

path of the arb filename, if unspecified, creates a temporary file
(nom du fichier arb généré (avec extension). Si non renseigné, un fichier temporaire.)

tab_filename

[mandatory] path of the tab_filename.
([obligatoire]nom du fichier .tab (avec extension).)

rda_filename

rda file name (with .rda extension)
nom du fichier rda (avec extension)

hst_filename

Apriori file name, syntax detailed below, Example : hst_filename = "path_to_file/apriori.hst"
(fichier(s) d'informations a priori. Voir ci-dessous pour la syntaxe. Exemple : hst_filename = "path_to_file/apriori.hst")

explanatory_vars

[mandatory] Explanatory vars in a vector Example : c("CJ", "A21") for the tabular CJ x A21
([obligatoire] variables catégorielles, sous forme de vecteur. Exemple : c("CJ", "A21") pour le premier Pour un tableau croisant CJ x A21)

value

Colname for response variable in the tabular "<freq>". For frequency table
(Nom de la variable de réponse dans le tableau "<freq>". Permet de tariter les tableaux de fréquence)

safety_rules

[mandatory] Rules for primary suppression with Argus syntax, if the primary suppression has been dealt with an apriori file specify manual safety range :"MAN(10)" for example.
([obligatoire] Règle(s) de secret primaire. Chaîne de caractères en syntaxe batch Tau-Argus. Si le secret primaire a été traité dans un fichier d'apriori : utiliser "MAN(10)")

suppress

[mandatory] Algortihm for secondary suppression (Tau-Argus batch syntax), and the parameters for it.
([obligatoire] Algorithme de gestion du secret secondaire (syntaxe batch de Tau-Argus), ainsi que les potentiels paramètres associés)

output_names

output file name
(nom du fichier en sortie.)

output_type

Type of the output file (Argus codification) By default "2" (csv for pivot-table). For SBS files use "4"
(Format des fichiers en sortie (codification Tau-Argus). Valeur par défaut du package : "2" (csv for pivot-table). Pour le format SBS utiliser "4")

output_options

Additionnal parameter for the output, by default : code"AS+" (print Status). To specify no options : "".
(Options supplémentaires des fichiers en sortie. Valeur par défaut du package : "AS+" (affichage du statut). Pour ne spécifier aucune option, "".)

gointeractive

Boolean, if TRUE will open a Tau-Argus window and launch the batch in it (FALSE by default).
(Possibilité de lancer le batch depuis le menu de Tau-Argus (FALSE par défaut).)

Value

A list containing two elements : the arb file name and the output name (usefull if the name is generated randomly)

(Une liste de deux éléments : le nom du fichier arb, le nom fichiers en sortie (utile pour récupérer les noms générés aléatoirement).)

Syntaxe

For tabular data, this function secures only one table at a time#' Unless specific details, use syntax from Tau-Argus manual
Sauf mention contraire, utiliser la syntaxe mentionnée dans la documentation de Tau-Argus.

Syntax pour suppress : First parameter is the table number, in this case it will always be 1

(Pour les données tabulées la fonction traite un tableau à la fois Syntaxe spéciale pour suppress : le premier paramètre dans la syntaxe Tau-Argus est le numéro du tableau. Dans le cas de cette fonction ce sera toujours 1)

Informations hst_filename

It's possible to add an apriori file (.hst) for a tabular, it can be generated by the table_rda() function.

Other options are not mandatory. To modify default value use a vector with the .hst file path as first element and then complete with those parameters : sep for the separator, ignore_err for IgnoreError and exp_triv for ExpandTrivial.

(Il est possible de fournir un fichier apriori (.hst) pour un tableau, il peut être fourni par la fonction table_rda() Exemple : hst_filename = "mon_chemin/apriori.hst"

Les options supplémentaires sont facultatives. Pour modifier les valeurs par défaut, passer une liste ayant comme premier élément le(s) fichier(s) hst et compléter avec les éléments portant les noms sep pour le séparateur, ignore_err pour IgnoreError et exp_triv pour ExpandTrivial. Comme pour les noms de fichiers, spécifier une seule valeur par paramètre ou autant de valeurs que de tabulations.)

Examples

if (FALSE) {
# creating arb file
infos_arb <- tab_arb(
  arb_filename = "path/tab1.arb",
  tab_filename = "path/tab1.tab",
  rda_filename = "path/tab1.rda",
  explanatory_vars = c("REGION", "CJ"),
  value = "CA",
  safety_rules = "NK(1,85)|FREQ(3,10)",
  suppress = "GH(1,100)",
  output_names = "path/tab1.csv",
  output_options = "AS+SE+",
  output_type = "4"
)
# show the content of the file in console
# visualisation du contenu du fichier dans la console
file.show(infos_arb$arb_filename, pager = "console")
}