All in once for tabular

All in once for tabular

Usage

tab_rtauargus(
  tabular,
  files_name = NULL,
  dir_name = NULL,
  explanatory_vars,
  totcode = getOption("rtauargus.totcode"),
  hrc = NULL,
  secret_var = NULL,
  cost_var = NULL,
  value = "value",
  freq = "freq",
  maxscore = NULL,
  safety_rules = "MAN(10)",
  suppress = "MOD(1,5,1,0,0)",
  show_batch_console = FALSE,
  output_type = 4,
  output_options = "",
  unif_labels = TRUE,
  ...
)

Arguments

tabular

[mandatory] data.frame which contains the tabulated data and an additional boolean variable that indicates the primary secret of type boolean
([obligatoire] data.frame contenant les données tabulées et une variable supplémentaire indiquant le secret primaire de type booléen.)

files_name

string used to name all the files needed to process. All files will have the same name, only their extension will be different.

dir_name

string indicated the path of the directory in which to save all the files (.rda, .hst, .txt, .arb, .csv) generated by the function.

explanatory_vars

[mandatory] Vector of explanatory variables
[obligatoire] Variables catégorielles, sous forme de vecteurs
Example : c("A21", "TREFF", "REG") for a table crossing A21 x TREFF x REG

totcode

Code(s) which represent the total of a categorical variable (see section 'Specific parameters' for this parameter's syntax). If unspecified for a variable(neither by default nor explicitly) it will be set to rtauargus.totcode.
(Code(s) pour le total d'une variable catégorielle (voir section 'Specific parameters' pour la syntaxe de ce paramètre). Les variables non spécifiées (ni par défaut, ni explicitement) se verront attribuer la valeur de rtauargus.totcode.)

hrc

Informations of hierarchical variables (see section 'Hierarchical variables').
(Informations sur les variables hiérarchiques (voir section 'Hierarchical variables').)

secret_var

Boolean variable which give the primary secret : equal to "TRUE" if a cell is concerned by the primary secret,"FALSE" otherwise. will be exported in the apriori file
(Variable indiquant le secret primaire de type booléen: prend la valeur "TRUE" quand les cellules du tableau doivent être masquées par le secret primaire, "FALSE" sinon. Permet de créer un fichier d'apriori)

cost_var

Numeric variable allow to change the cost suppression of a cell for secondary suppression, it's the value of the cell by default, can be specified for each cell, fill with NA if the cost doesn't need to be changed for all cells
(Variable numeric qui permet de changer la coût de suppression d'une cellule, pris en compte dans les algorithmes de secret secondaire.Par défaut le coût correspond à la valeur de la cellule. peut être spécifié pour chacune des cellules, peut contenir des NA pour les coûts que l'on ne souhaite pas modifier.)

value

Name of the column containing the value of the cells.
(Nom de la colonne contenant la valeur des cellules)

freq

Name of the column containing the cell frequency.
(Nom de la colonne contenant les effectifs pour une cellule)

maxscore

Name of the column containing, the value of the largest contributor of a cell.
(Nom de la colonne contenant la valeur du plus gros contributeur d'une cellule)

safety_rules

[mandatory] Rules for primary suppression with Argus syntax, if the primary suppression has been dealt with an apriori file specify manual safety range :"MAN(10)" for example.
([obligatoire] Règle(s) de secret primaire. Chaîne de caractères en syntaxe batch Tau-Argus. Si le secret primaire a été traité dans un fichier d'apriori : utiliser "MAN(10)")

suppress

[mandatory] Algortihm for secondary suppression (Tau-Argus batch syntax), and the parameters for it.
([obligatoire] Algorithme de gestion du secret secondaire (syntaxe batch de Tau-Argus), ainsi que les potentiels paramètres associés)

show_batch_console

to display the batch progress in the console.
(pour afficher le déroulement du batch dans la console.)

output_type

Type of the output file (Argus codification) By default "2" (csv for pivot-table). For SBS files use "4"
(Format des fichiers en sortie (codification Tau-Argus). Valeur par défaut du package : "2" (csv for pivot-table). Pour le format SBS utiliser "4")

output_options

Additionnal parameter for the output, by default : code"AS+" (print Status). To specify no options : "".
(Options supplémentaires des fichiers en sortie. Valeur par défaut du package : "AS+" (affichage du statut). Pour ne spécifier aucune option, "".)

unif_labels

boolean, if explanatory variables have to be standardized

...

any parameter of the tab_rda, tab_arb or run_arb functions, relevant for the treatment of tabular.

Value

If output_type equals to 4, then the original tabular is returned with a new column called Status, indicating the status of the cell coming from Tau-Argus : "A" for a primary secret due to frequency rule, "B" for a primary secret due to dominance rule, "D" for secondary secret and "V" for no secret cell.

If output_type doesn't equal to 4, then the raw result from tau-argus is returned.

Standardization of explanatory variables and hierarchies

The boolean argument unif_labels is useful to prevent some common errors in using Tau-Argus. Indeed, Tau-Argus needs that, within a same level of a hierarchy, the labels have the same number of characters. When the argument is set to TRUE, tab_rtauargus standardizes the explanatory variables to prevent this issue. Hierarchical explanatory variables (explanatory variables associated to a hrc file) are then modified in the tabular data and an another hrc file is created to be relevant with the tabular. In the output, these modifications are removed.

Examples

if (FALSE) {
library(dplyr)
data(turnover_act_size)

# Prepare data with primary secret ----
turnover_act_size <- turnover_act_size %>%
  mutate(
    is_secret_freq = N_OBS > 0 & N_OBS < 3,
    is_secret_dom = ifelse(MAX == 0, FALSE, MAX/TOT>0.85),
    is_secret_prim = is_secret_freq | is_secret_dom
  )

# Make hrc file of business sectors ----
data(activity_corr_table)
hrc_file_activity <- activity_corr_table %>%
  write_hrc2(file_name = "hrc/activity")

# Compute the secondary secret ----
options(
  rtauargus.tauargus_exe =
    "Y:/Logiciels/TauArgus/TauArgus4.2.2b1/TauArgus.exe"
)

res <- tab_rtauargus(
  tabular = turnover_act_size,
  files_name = "turn_act_size",
  dir_name = "tauargus_files",
  explanatory_vars = c("ACTIVITY", "SIZE"),
  hrc = c(ACTIVITY = hrc_file_activity),
  totcode = c(ACTIVITY = "Total", SIZE = "Total"),
  secret_var = "is_secret_prim",
  value = "TOT",
  freq = "N_OBS",
  verbose = FALSE
)
}